Tests/LibWeb/Text/expected/Editing/insertParagraph-at-host-start.txt

after enter: "<div><br></div><div>line1<br>line2</div>"
after undo: "line1<br>line2"
